+++ tests/packerlicious/test_post_processor_shell_local.py
@@ -10,4 +10,4 @@ class TestShellLocalPostProcessor(object):
 
         with pytest.raises(ValueError) as excinfo:
             b.to_dict()
-        assert u'one of the following must be specified: inline, script, scripts' in str(excinfo.value)
+        assert u'one of the following must be specified: command, inline, script, scripts' in str(excinfo.value)